Add a default secrets option (#12)
* Add a default secret option

This adds a `default-secret-name` config which when provided,
identifies a k8s secret in the cluster from which to
retrieve credentials. The secret is expected to have the
following keys and they are non-configurable.
```
    buildkite-agent-token: <token>
    git-credentials: <credentials>
    ssh-key: <sshkey>
```
The idea is to avoid the need to manually provide the values in the
pipeline in the cases where the user has control over secret creation.