Change pod creation timeout to unique exit code (#35)
* Change pod creation timeout to unique exit code

If the pod isn't created before the timeout is reached, this adds a
unique exit code so users can automatically retry in this case. Without
this it exits 1 which is difficult to differentiate between a real job
error in most cases. Ideally buildkite would show a timeout in this case
since this timeout is derived from that, but it's seems to race that

* Update hooks/command

Co-authored-by: Ifeanyi Ubah <ify1992@yahoo.com>

* Update hooks/command

Co-authored-by: Ifeanyi Ubah <ify1992@yahoo.com>

* Use other syntax]

Co-authored-by: Ifeanyi Ubah <ify1992@yahoo.com>